Add accessors for TV show metadata
  - Add method for determining whether a track is a TV show
  - Add accessors for season_number and episode_number
  - Update "iTunes Library.xml" fixture to include a TV show.
    (This data was extracted from another "iTunes library.xml"
    containing a TV episode purchased from iTunes.)

Note: The XML file stores the episode number in an oddly-named
field called "Episode Order".  It stores the season number in a
field called "Season".  The accessors in this commit use more
friendly names to refer to these fields (i.e., "Episode Number"
and "Season Number" respectively).  These also happen to be the
names used in the iTunes UI:

  http://skitch.com/jasonrudolph/frebh/itunes-tv-show-metadata